% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined_cutoff_percent.R
\name{Combined_cutoff_percent}
\alias{Combined_cutoff_percent}
\title{Identification of Careless Reponse Percentages Using Standard Deviation and Time per Item}
\usage{
Combined_cutoff_percent(
  data,
  SD.cutoff,
  TPI.cutoff,
  ttc.colnames,
  number.items,
  mandatory.response,
  item.colnames,
  ID.colname
)
}
\arguments{
\item{data}{dataframe to be analyzed.}

\item{SD.cutoff}{numeric cutoff value for Standard Deviation.}

\item{TPI.cutoff}{numeric cutoff value for Time per Item.}

\item{ttc.colnames}{vector of column names of "Start Time" and "End Time" to calculate time to complete, also can be "Completion Time" if already calculated.}

\item{number.items}{integer, number of items per assessment.}

\item{mandatory.response}{logical value based on whether response to items in each assessment were mandatory to complete.}

\item{item.colnames}{vector of column names of all items/questions to be used to calculate item score Standard Deviation and Longstring responses.}

\item{ID.colname}{character string of column name for ID of assessment.}
}
\value{
The item \code{"ttc.colnames"} must be the names of columns, corresponding to Start Time and End Time, ordered \code{ttc.colnames = c("StartTime", "EndTime")} start time first, follwed by end time. If the data includes assessment duration, then list the column name that corresponds with assessment completion: \code{ttc.colnames = "SurveyDuration"}

The item \code{"mandatory.response"} should be \strong{\code{TRUE}} if participants were required to complete all items per assessment. This item should be \strong{\code{FALSE}} if not all items were required to be completed per assessment.

If there is variability in the items that are asked, mark the item \code{"mandatory.response"} as \strong{\code{FALSE}} and ensure that \code{"item.colnames"} includes all items.
}
\description{
\code{Combined_cutoff_percent} returns dataframe of percent of assessment per ID flagged by two cutoff values: Standard Deviation per assessment and Time per Item.
}
\details{
This function creates the dataframe that includes the ID and the percentage of assessments that met the cutoff criteria for \emph{Standard Deviation per assessment} as well as \emph{Time per Item}. 
The percentage is calculated as number of assessments that met both criteria by the total number of assessments for the individual ID.
If an assessment has a Standard Deviation \strong{less than or equal to} the cutoff value and a Time per Item \strong{less than or equal to} the cutoff value, it will be flagged and placed in the dataframe. 
This function uses the Boolean logic \emph{"AND"} in order to identify assessments that meet both criteria.
}
\references{
Jaso, B.A., Kraus, N.I., Heller, A.S. (2020) \emph{Methods to automatically identify careless responding in ecological momentary assessment research: from post-hoc analyses to real-time data monitoring.}
}
\seealso{
\code{\link{Combined_cutoff_percent}} for a similar fuction, presenting only ID and the percent of assessments that were flagged by the criteria.

See the following functions for more information on Careless Response Identification in EMA: \code{\link{flagging_df}}, \code{\link{flagging_plots}}, \code{\link{longstringr}}, \code{\link{SD_cutoff}}, and \code{\link{TPI_cutoff}}
}
