% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMAeval_Data.R
\docType{data}
\name{EMAeval_Data}
\alias{EMAeval_Data}
\title{Example data set for the EMAeval Package}
\format{
A dataframe with 200 rows and 11 variables:
\describe {
    \item{ID}{identification number given to all 4 participants.}
    \item{StartDate}{date time when the assessment began.}
    \item{EndDate}{date time when the assessment ended.}
    \item{Q#}{individual items asked of participants for each assessment, numbered 1-8.}
}
}
\source{
This data set is ficticious data only to demonstrate the uses of the R package EMAeval. Developed by Noah Kraus.
}
\usage{
data(EMAeval_Data)
}
\description{
This is a ficticious data set that has 4 "participants" who were sampled once every day, for 50 straight days. There are responses to 8 different items (Q1, Q2, Q3, etc.).
This data can be used to help users understand the use of the EMAeval R Package.
}
\keyword{datasets}
